+++ bugsnag-spring/src/test/java/com/bugsnag/testapp/springboot/TestController.java
@@ -8,7 +8,6 @@ import com.bugsnag.callbacks.Callback;
 
 import org.springframework.beans.TypeMismatchException;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskScheduler;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 
@@ -19,9 +18,6 @@ public class TestController {
 
     @Autowired
     private Bugsnag bugsnag;
-
-    @Autowired
-    private ThreadPoolTaskScheduler scheduler;
 
     /**
      * Throw a runtime exception
@@ -78,21 +74,5 @@ public class TestController {
                 }
             });
         }
-    }
-
-    /**
-     * Schedule a task to throw an unhandled exception
-     */
-    @RequestMapping("/execute-scheduled-task")
-    public void executeScheduledTask() throws ExecutionException, InterruptedException {
-        Runnable exampleRunnable = new Runnable() {
-            @Override
-            public void run() {
-                throw new RuntimeException("Scheduled test");
-            }
-        };
-
-        // Run the task now and wait for it to finish
-        scheduler.submit(exampleRunnable).get();
     }
 }
